Move create_history_template out of galaxy.util.
Stuff in galaxy.util definitely shouldn't import galaxy.web and ideally wouldn't operate on models or use trans/app either. This file was only used in one place so it likely belongs sitting right next to that code.
