+++ lib/galaxy/webapps/galaxy/controllers/_create_history_template.py
