+++ lib/galaxy/webapps/galaxy/controllers/history.py
@@ -15,7 +15,6 @@ from galaxy.model.item_attrs import (
     UsesItemRatings
 )
 from galaxy.util import listify, Params, parse_int, sanitize_text
-from galaxy.util.create_history_template import render_item
 from galaxy.util.odict import odict
 from galaxy.web import url_for
 from galaxy.web.base.controller import (
@@ -29,6 +28,7 @@ from galaxy.web.base.controller import (
     WARNING,
 )
 from galaxy.web.framework.helpers import grids, iff, time_ago
+from ._create_history_template import render_item
 
 
 log = logging.getLogger(__name__)
